feat(CounterField): add new helpers file for clamValue and solved id and uesmemo issue
