+++ src/helpers/index.ts
@@ -1,1 +1,2 @@
 export { default as allowOnlyDigits } from "./allowOnlyDigits";
+export { default as clampValue } from "./clampValue";
